airflow_dags/requirements_airflow.txt
apache-airflow
boto3
numpy
pandas
scikit-learn
scipy
xgboost
openpyxl
psycopg2-binary